Refactor dialogs utils and fix origination fee in new loan dialog
